#5 [FEAT] 유틸리티 함수 수정, 분리 및 회원가입 연동 (주소관련 오류 수정해야함)
